requirements.txt
pandas
chromadb
huggingface-hub
python-dotenv
streamlit
matplotlib
langchain
langchain-community
langchain-huggingface
langchain-chroma
pydantic
pydantic-core
wordcloud
sentence-transformers
requests
pytest